% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FD_df.R
\name{FD_df}
\alias{FD_df}
\title{Transforming a time-bins x life-mode dataframe into a functional diversity dataframe}
\usage{
FD_df(df, subcat)
}
\arguments{
\item{df:}{\code{dataframe}. Input dataframe with row names indicating time bins and column names indicating each life-mode or ecocode. The \code{df} values are estimates of the abundance of each life-mode at each time bin.}

\item{subcat:}{\code{numeric vector}. The number of subcategories for each trait.Its length must be equal that the number of traits of each life-mode.}
}
\value{
This function returns a functional diversity dataframe with an extension of the traits defined by the ecocode
        and the corresponding estimation of abundance or biomass for each time bin.
}
\description{
This function transforms an input dataframe \code{df} into a functional diversity dataframe.
}
\details{
An \emph{ecocode} is defined by a numeric sequence of traits as integers (e.g., 11211234); hence, the length of the \code{subcat} argument must be eight and its values cannot be numerically lower than the respective values in the ecocode's traits.
}
\examples{
View(data0) #internal dataset

data1 <- Fd_df(df=data3char)

View(data1) #functional diversity dataframe



}
\author{
Simon P. Castillo \email{spcastil@uc.cl} & Felipe Opazo-Mella
}
