% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FD_nullPlot.R
\name{FD_nullPlot}
\alias{FD_nullPlot}
\title{Plot the temporal trajectory in functional diversity estimated from empirical data and null models}
\usage{
FD_nullPlot(
  d.rao,
  d.FD,
  sumRao,
  sumFD,
  plotRao = TRUE,
  saveRao = TRUE,
  plotFRic = TRUE,
  saveFRic = TRUE
)
}
\arguments{
\item{d.rao:}{a dataframe obtained after running the function \code{\link{FD_nullSummary}}.}

\item{d.FD:}{a dataframe obtained after running the function \code{\link{FD_nullSummary}}.}

\item{sumRao:}{a dataframe obtained after running the function \code{\link{FD_Rao}}.}

\item{sumFD:}{a dataframe obtained after running the function \code{\link{FD_metrics}}.}

\item{plotRao:}{\code{TRUE} or \code{FALSE}. Do wou want to plot the Rao's, Simpson's and functional redundancy estimates with the null models. Default \code{TRUE}.}

\item{saveRao:}{\code{TRUE} or \code{FALSE}. Do wou want to save the Rao's, Simpson's and functional redundancy estimates with the null models. Default \code{TRUE}.}

\item{plotFRic:}{\code{TRUE} or \code{FALSE}. Do wou want to plot the functional richness (FRic) estimates with the null models? Default \code{TRUE}.}

\item{saveFRic:}{\code{TRUE} or \code{FALSE}. Do wou want to save the Rao's, Simpson's and functional redundancy estimates with the null models. Default \code{TRUE}.}
}
\value{
The function returns a named list with each plot made with \code{\link[ggplot2:ggplot]{ggplot2}} and  prints in the plot window (if \code{plotRao} and/or \code{plotFRic} are \code{TRUE}); and will save the corresponding plots in \code{png} format in the \code{wd}.
}
\description{
This function plot the summary data returned by the function \code{\link{FD_nullSummary}}. This function can be called directly from \code{\link{FD_nullcom}} and from \code{\link{FD_nullsummary}}.
}
\examples{
FD_plotNull(d.rao = summarynullRao, d.FD = summarynullFD, plotRao = TRUE, saveRao = TRUE, plotFRic = TRUE, saveFRic = TRUE)



}
\seealso{
\code{\link[ggplot2:ggplot]{ggplot2}}
}
\author{
Simon P. Castillo \email{spcastil@uc.cl}, Felipe Opazo-Mella, Fabio Labra and Pablo Marquet
}
