% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FD_nullSummary.R
\name{FD_nullSummary}
\alias{FD_nullSummary}
\title{Summarise the null communities' functional diversity}
\usage{
FD_nullSummary(nullrao, nullFD, nNull, plot.null = TRUE)
}
\arguments{
\item{nullrao:}{dataframe obtained after computing the function \code{\link{FD_nullCom}}.}

\item{nullFD:}{dataframe obtained after computing the function \code{\link{FD_nullCom}}.}

\item{nNull:}{numeric. The number of null replicates. This must be the same that those in \code{\link{FD_nullCom}}.}
}
\value{

}
\description{
This function summarises the dataframes returned by the function \code{\link{FD_nullCom}}. This function can be called directly from \code{\link{FD_nullCom}}.
}
\examples{
FD_nullSummary(nullrao = null_rao, nullFD = null_FDmetrics, nNull = 3, plot.null = FALSE)

View(summarynullrao)

View(summarynullFD)



}
\author{
Simon P. Castillo \email{spcastil@uc.cl}, Felipe Opazo-Mella, Fabio Labra and Pablo Marquet
}
