% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FD_plotFS.R
\name{FD_plotFS}
\alias{FD_plotFS}
\title{Plot the functional space form empirical data}
\usage{
FD_plotFS(
  dfFD,
  sumFD,
  sumRao,
  plot.window = TRUE,
  cols = viridis,
  save.plot = TRUE,
  add.labs = FALSE,
  ext = ".png"
)
}
\arguments{
\item{dfFD:}{dataframe. A functional diversity dataframe created with the function \code{\link{FD_df}}.}

\item{sumRao:}{a dataframe obtained after running the function \code{\link{FD_Rao}}.}

\item{sumFD:}{a dataframe obtained after running the function \code{\link{FD_metrics}}.}

\item{plot.window:}{\code{TRUE} or \code{FALSE} if the plots will appear in the Rstudio plot window.Default \code{TRUE}.}

\item{save.plot:}{\code{TRUE} or \code{FALSE}. Default \code{TRUE}.}

\item{add.labs:}{\code{TRUE} or \code{FALSE}. It adds the label corresponding to each life-mode or ecocode. Default \code{FALSE}.}

\item{ext:}{extension of the plot to be saved. Only extenssion admitted by \code{\link[ggplot2]{ggsave}} (e.g., \code{".svg"}, \code{".png"}). Default \code{".png"}.}
}
\value{
The function returns a list with a plot for each time bin. If \code{plot.window} is \code{TRUE} the plot will appear in the IDE plot window. If \code{save.plot} is \code{TRUE} plots will be saved with a \code{".png"} extension in your \code{wd}.
}
\description{
Plot the functional space for each time bin as paired coordinates PC1 vs PC2 and PC1 vs PC3.
}
\examples{
View(FDsummary)

FD_plotFS(FDsummary, plot.window=TRUE,save.plot= TRUE,add.labs=FALSE, ext=".png" )

plotsFS <- FD_plotFS(FDsummary, plot.window=TRUE,save.plot= TRUE,add.labs=FALSE, ext=".png" ) #
#plotsFS is a list

}
\seealso{
\code{\link{FD_metrics}}, \code{\link[ggplot2]{ggsave}}.
}
\author{
Simon P. Castillo \email{spcastil@uc.cl}, Felipe Opazo-Mella, Fabio Labra and Pablo Marquet
}
