% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FD_plotRao.R
\name{FD_plotRao}
\alias{FD_plotRao}
\title{Plot functional metrics over time from empirical data}
\usage{
FD_plotRao(df, save.plot = TRUE, ext = ".png")
}
\arguments{
\item{df:}{dataframe obtained after running \code{\link{FD_Rao}}.}

\item{save.plot:}{\code{TRUE} or \code{FALSE}. Default \code{TRUE}.}

\item{ext:}{extension for the saved figures admitted by \code{\link[ggplot2]{ggsave}} (e.g., \code{".png"}, \code{".svg"}).}
}
\value{
Plots of Rao's quadratic entropy, functional redundancy and Simpson's richness estimates over time (raw and normalised). If \code{save.plot} is \code{TRUE}, the plots will be saved too.
If the function is assigned it returns a list with two plots \code{plot1} and \code{plot2}, with the raw and normalised plots, respectively.
}
\description{
Plot functional metrics (Simpson, Rao's quadratic Entropy and Funcional redundancy) over time from empirical data.
}
\details{
The normalisation is performed within each variable taking the minimum and maximum values.
}
\examples{
View(RaoSummary) # dataframe returned by DF_Rao

FD_plotRao(RaoSummary, save.plot=TRUE, ext=".png")
}
\seealso{
\code{\link{FD_Rao}}.
}
\author{
Simon P. Castillo \email{spcastil@uc.cl} & Felipe Opazo-Mella
}
