% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FD_Rao.R
\name{FD_Rao}
\alias{FD_Rao}
\title{Functional metrics (Simpson, Rao's quadratic Entropy and Funcional redundancy)}
\usage{
FD_Rao(df)
}
\arguments{
\item{df:}{is a functional diversity dataframe obtained by using the function \code{\link{FD_df}}.}
}
\value{
This function returns a dataframe with the Rao's complexity, Simpson's diversity, and functional redundancy estimates.
}
\description{
This functions calculates the Rao's quadratic entropy, Simpson's index and functional redundancy for each time bin along the temporal series.
}
\note{
This function uses \code{\link[SYNCSA]{rao.diversity}} package for each time bin as a community.
}
\examples{
RaoSummary <- FD_rao(df=data1)

View(RaoSummary)
}
\references{
Rao, C.R. (1982). Diversity and dissimilarity coefficients: a unified approach. Theoretical Population Biology, 21, 24:43.
}
\seealso{
\code{\link[SYNCSA]{rao.diversity}}, \code{\link[FD]{gowdis}}
}
\author{
Simon P. Castillo \email{spcastil@uc.cl} & Felipe Opazo-Mella
}
