% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FD_rarefaction.R
\name{FD_rarefaction}
\alias{FD_rarefaction}
\title{Compute rarefaction analyses}
\usage{
FD_rarefaction(df, times = 500)
}
\arguments{
\item{df:}{a dataframe containing the database of ocurrences at a given taxonomic level. It must match column names of \code{FD_data0}.}

\item{time:}{numeric. The number of iterations to run. Default 500.}
}
\value{
The function saves three dataframes in your working directory name "Diversity.csv", "EcoGenera.csv" and "EcoOccurrences.csv".

"Diversity.csv": stores diversity and ecodiversity estimates by time bin.

"EcoOccurrences.csv": stores the number of ocurrences by ecocode over time.

"EcoGenera.csv": stores the number of genera by ecocode over time.
}
\description{
Compute rarefaction analyses
}
\examples{
FD_rarefaction(data=FD_data0, times=500)



}
\author{
Simon P. Castillo \email{spcastil@uc.cl}, Felipe Opazo-Mella, Fabio Labra and Pablo Marquet
}
