% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FD_data.R
\docType{data}
\name{FDSeries-Datasets}
\alias{FDSeries-Datasets}
\title{Internal datasets}
\usage{
View(df)
}
\arguments{
\item{df:}{*dataframe*. Is one of the two dataframes in this package 'data3char' or 'data8char'.}
}
\description{
These toy datasets contain the information of abundaance over time for different sets of functional traits.The dataset 'datachar3' have life-mode or ecocodes with with three characters each one; while in the dataset
'data8char', each ecocode have eight characters.
}
\examples{
 ncol(data3char) #Number of time bins
 nrow(data3char) #Number of life-modes or ecocodes

 ncol(data8char) #Number of time bins
 nrow(data8char) #Number of life-modes or ecocodes


}
\author{
Simon P. Castillo \email{spcastil@uc.cl} & Felipe Opazo-Mella
}
\keyword{datasets}
