% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_FD.R
\name{ind_FD}
\alias{ind_FD}
\title{Internal use}
\source{
This is an edited version from Villeger: \url{http://villeger.sebastien.free.fr/Rscripts.html}
}
\usage{
ind_FD(traits, abundances)
}
\value{
dataframe
}
\description{
For internal use. Computes functional diversity metrics for a single community (one time bin). Edited from original source \url{http://villeger.sebastien.free.fr/Rscripts.html} See reference below.
}
\references{
Villéger, S., Mason, N. W., & Mouillot, D. (2008). New multidimensional functional diversity indices for a multifaceted framework in functional ecology. Ecology, 89(8), 2290-2301.

Bellwood, D. R., Wainwright, P. C., Fulton, C. J., & Hoey, A. S. (2006). Functional versatility supports coral reef biodiversity. Proceedings of the Royal Society B: Biological Sciences, 273(1582), 101-107.
}
\seealso{
\code{\link{FD_metrics}}, \code{\link[FD:FD-package]{FD}}
}
