% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullmodel_FD.R
\name{nullmodel_FD}
\alias{nullmodel_FD}
\title{Internal use}
\usage{
nullmodel_FD(com, type = "all")
}
\arguments{
\item{com:}{a timexlife-mode dataframe. time bins in rows and life-mode in columns.}

\item{type:}{Type of resampling. \code{"col-wise"} relocates abundances within a life-mode, \code{"row-wise"} relocates abundances wihtin a time bin, and \code{"all"} relocate abundances across time bins and life-modes.}
}
\value{
This function returns a resampled data.frame (null community).
}
\description{
This function is for internal use; it computed the null models. There are three options based on the resampling routine.
}
\examples{
nullcommunity <- nullmodel_FD(com=data0, type="all")



}
\author{
Simon P. Castillo \email{spcastil@uc.cl} & Felipe Opazo-Mella
}
